% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agalmar.R
\docType{package}
\name{agalmar}
\alias{agalmar}
\alias{agalmar-package}
\title{agalmar: A Collection of Tools for Analyzing Agalma Expression Projects}
\description{
The agalmar package contains a collection of tools for analyzing 
expression and phylogenetic data exported from Agalma.
See https://bitbucket.org/caseywdunn/agalma
}
