% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{apply_normalizations}
\alias{apply_normalizations}
\title{Takes a DGEList object, and returns a matrix of normalized counts}
\usage{
apply_normalizations(dge)
}
\arguments{
\item{dge}{edgeR DGEList object, to which normalizations have been applied}
}
\value{
matrix of normalized counts
}
\description{
Takes a DGEList object, and returns a matrix of normalized counts
}

