% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{create_DESeq2,Expression,formula-method}
\alias{create_DESeq2,Expression,formula-method}
\title{Creates a DESeqDataSet object}
\usage{
\S4method{create_DESeq2}{Expression,formula}(object, design)
}
\arguments{
\item{object}{An Expression object}

\item{design}{a formula that explains the project design}
}
\value{
A DESeqDataSet object
}
\description{
Creates a DESeqDataSet object
}
