% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_DESeq2}
\alias{create_DESeq2}
\title{Creates a DESeqDataSet object}
\usage{
create_DESeq2(object, design)
}
\arguments{
\item{object}{An Expression object}

\item{design}{a formula that explains the project design}
}
\value{
A DESeqDataSet object
}
\description{
Creates a DESeqDataSet object
}
