% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dataframe_to_node_labels}
\alias{dataframe_to_node_labels}
\title{Converts data frame of node annotations to NHX node labels}
\usage{
dataframe_to_node_labels(df)
}
\arguments{
\item{df}{Data frame
return character A vector of node label names}
}
\description{
Converts data frame of node annotations to NHX node labels
}
