% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{decompose_orthologs}
\alias{decompose_orthologs}
\title{Decomposes a gene tree into a list of subtrees that have no duplication events. Assumes notung
style node annotations.}
\usage{
decompose_orthologs(phy)
}
\arguments{
\item{phy}{The tree, as an ape phylo object}
}
\value{
The subtrees as a list of phylo object
}
\description{
Decomposes a gene tree into a list of subtrees that have no duplication events. Assumes notung
style node annotations.
}

