% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{drop.tip.nhx}
\alias{drop.tip.nhx}
\title{Drops tips from an NHX tree, similar to how ape::drop.tip 
drops tips from a phylo object}
\usage{
drop.tip.nhx(nhx, tip, test = FALSE)
}
\arguments{
\item{nhx}{A ggtree nhx object}

\item{tip}{A vector of mode numeric or character specifying the tips to delete}

\item{test}{A logical to designate whether to rotate some internal nodes for more robust testing}
}
\value{
A ggtree::nhx object
}
\description{
Drops tips from an NHX tree, similar to how ape::drop.tip 
drops tips from a phylo object
}

