% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Expression}
\alias{Expression}
\title{Construct an Expression object from list of experiment data and metadata 
provided by Agalma.}
\usage{
Expression(data_list)
}
\arguments{
\item{data_list}{A list containing the expression data}
}
\value{
An Expression object
}
\description{
Construct an Expression object from list of experiment data and metadata 
provided by Agalma.
}
