% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_lane}
\alias{get_lane}
\title{Get lane from Illumina fastq sequence header}
\usage{
get_lane(header)
}
\arguments{
\item{header}{character strings}
}
\value{
lane character strings
}
\description{
Get lane from Illumina fastq sequence header
}
