% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_run}
\alias{get_run}
\title{Get run from Illumina fastq sequence header}
\usage{
get_run(header)
}
\arguments{
\item{header}{character strings}
}
\value{
run character strings
}
\description{
Get run from Illumina fastq sequence header
}
