% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_tip_info}
\alias{get_tip_info}
\title{Parse species and sequence id from a phy}
\usage{
get_tip_info(phy)
}
\arguments{
\item{phy}{The tree, as an ape phylo object. Tip names must have `species@id` format.}
}
\value{
Dataframe with one row per tip, a species column, and an id column
}
\description{
Parse species and sequence id from a phy
}

