% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{janedoe}
\alias{janedoe}
\title{Expression data for different tissues from an animal}
\format{A list parsed from a json that was exported 
from agalma with 
https://bitbucket.org/caseywdunn/agalma/src/master/scripts/agalma-export-expression}
\usage{
janedoe
}
\description{
Expression data for different tissues from an animal
}
\keyword{datasets}

