% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{nhx_label_to_list}
\alias{nhx_label_to_list}
\title{Converts a NHX node label string to a list of named values}
\usage{
nhx_label_to_list(label)
}
\arguments{
\item{label}{Character string}
}
\value{
list A named list of values
}
\description{
Converts a NHX node label string to a list of named values
}
\examples{
nhx_label_to_list( "[&&NHX:Ev=S:S=58:ND=0]" )
}
