% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{node_support}
\alias{node_support}
\title{Parse support from node name}
\usage{
node_support(node_name)
}
\arguments{
\item{node_name}{A character string of format '83:N', where '83' is node support
and N indicates whether the node is a duplication or not}
}
\value{
Numeric indicating node support
}
\description{
Parse support from node name
}

