% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{species,Expression-method}
\alias{species,Expression-method}
\title{Get the species}
\usage{
\S4method{species}{Expression}(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
Species name
}
\description{
Get the species
}
