% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{summarize_libraries,Expression-method}
\alias{summarize_libraries,Expression-method}
\title{Summarize expression libraries}
\usage{
\S4method{summarize_libraries}{Expression}(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame that summarizes each library with expression data in the
Expression object
}
\description{
Summarize expression libraries
}
