% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_libraries}
\alias{summarize_libraries}
\title{Summarize expression libraries}
\usage{
summarize_libraries(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame that summarizes each library with expression data in the
Expression object
}
\description{
Summarize expression libraries
}

