% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_nodes}
\alias{summarize_nodes}
\title{Create a data frame with summary statistics for edges in a phyldog NHX tree}
\usage{
summarize_nodes(nhx)
}
\arguments{
\item{nhx}{A ggtree nhx object}
}
\value{
A data frame of node summary statistics
}
\description{
Create a data frame with summary statistics for edges in a phyldog NHX tree
}

