% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{summarize_reference,Expression-method}
\alias{summarize_reference,Expression-method}
\title{Summarize reference sequences}
\usage{
\S4method{summarize_reference}{Expression}(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame that summarizes the reference sequences in the
Expression object
}
\description{
Summarize reference sequences
}

