% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_reference}
\alias{summarize_reference}
\title{Summarize reference sequences}
\usage{
summarize_reference(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame that summarizes the reference sequences in the
Expression object
}
\description{
Summarize reference sequences
}
