% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\docType{methods}
\name{summary_frame,Expression-method}
\alias{summary_frame,Expression-method}
\title{Create a summary table.}
\usage{
\S4method{summary_frame}{Expression}(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame
}
\description{
Create a summary table.
}

