% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary_frame}
\alias{summary_frame}
\title{Summarize experiment.}
\usage{
summary_frame(object)
}
\arguments{
\item{object}{An Expression object}
}
\value{
A data frame that summarizes each sample in the experiment
}
\description{
Summarize experiment.
}

