% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary_libraries}
\alias{summary_libraries}
\title{Create a data frame with summary statistics for expression
libraries}
\usage{
summary_libraries(e)
}
\arguments{
\item{e}{A list of Expression objects}
}
\value{
A data frame of summary statistics
}
\description{
Create a data frame with summary statistics for expression
libraries
}
