% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary_references}
\alias{summary_references}
\title{Create a data frame with summary statistics for reference sequences}
\usage{
summary_references(e)
}
\arguments{
\item{e}{A list of Expression objects}
}
\value{
A data frame of summary statistics
}
\description{
Create a data frame with summary statistics for reference sequences
}

