% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{decompose_orthologs}
\alias{decompose_orthologs}
\title{Decomposes a gene tree into a list of subtrees that have no duplication events. Assumes notung
style node annotations.}
\usage{
decompose_orthologs(nhx)
}
\arguments{
\item{nhx}{The tree, as a treeio nhx object}
}
\value{
The subtrees as a list of ape::phylo object
}
\description{
Decomposes a gene tree into a list of subtrees that have no duplication events. Assumes notung
style node annotations.
}
