% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get_lane}
\alias{get_lane}
\title{Get lane from Illumina fastq sequence header}
\usage{
get_lane(header)
}
\arguments{
\item{Header}{character strings}
}
\value{
Lane character strings
}
\description{
Get lane from Illumina fastq sequence header
}

