% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{has_species}
\alias{has_species}
\title{Returns true if all species are in the tree tips}
\usage{
has_species(phy, species)
}
\arguments{
\item{phy}{The tree, as an ape phylo object}
}
\value{
logical
}
\description{
Returns true if all species are in the tree tips
}

