% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{node_support}
\alias{node_support}
\title{Parse support from node name}
\usage{
node_support(node_name)
}
\arguments{
\item{node_name}{A character string of format 'n38047889-N-83', where 'n38047889-N' is 
notung annotation and '83' is node support.}
}
\value{
Numeric indicating node support
}
\description{
Parse support from node name
}

