% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_edges}
\alias{summarize_edges}
\title{Create a data frame with summary statistics for edges in a phyldog NHX tree}
\usage{
summarize_edges(nhx, default_length_val = NA)
}
\arguments{
\item{nhx}{A treeio nhx object}

\item{default_length_val}{The default length of branches}
}
\value{
A data frame of edge summary statistics
}
\description{
Create a data frame with summary statistics for edges in a phyldog NHX tree
}

