% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_nodes}
\alias{summarize_nodes}
\title{Create a data frame with summary statistics for nodes in a phyldog NHX tree}
\usage{
summarize_nodes(nhx, default_length_val = NA)
}
\arguments{
\item{nhx}{A treeio nhx object}

\item{default_length_val}{The default length of branches, used to determine
if a node has descendant branches with default length}
}
\value{
A data frame of node summary statistics
}
\description{
Create a data frame with summary statistics for nodes in a phyldog NHX tree
}
